% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrFNN.R
\docType{data}
\name{LearnerRegrFNN}
\alias{LearnerRegrFNN}
\alias{mlr_learners_regr.fnn}
\title{Regression fnn Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerRegr}.}
\usage{
LearnerRegrFNN
}
\description{
A \link{LearnerRegr} for a regression fnn implemented in FNN::knn()] in package \CRANpkg{FNN}.
}
\keyword{datasets}
