% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learner.fnn-package}
\alias{mlr3learner.fnn}
\alias{mlr3learner.fnn-package}
\title{mlr3learner.fnn: knn learner from FNN package for mlr3}
\description{
knn learner from FNN package for mlr3.
}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

}
