% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learnerfnn-package}
\alias{mlr3learnerfnn}
\alias{mlr3learnerfnn-package}
\title{mlr3learnerfnn: knn learner from FNN package for mlr3}
\description{
knn learner from FNN package for mlr3.
}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

}
