% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifFNN.R
\docType{data}
\name{LearnerClassifFNN}
\alias{LearnerClassifFNN}
\alias{mlr_learners_classif.fnn}
\title{Classification fnn Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifFNN
}
\description{
A \link[mlr3:LearnerClassif]{mlr3::LearnerClassif} for a classification fnn implemented in \code{\link[FNN:knn]{FNN::knn()}} in package \CRANpkg{FNN}.
}
\keyword{datasets}
