% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.fnn-package}
\alias{mlr3learners.fnn}
\alias{mlr3learners.fnn-package}
\title{mlr3learners.fnn: knn learner from FNN package for mlr3}
\description{
knn learner from FNN package for mlr3.
}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

Authors:
\itemize{
  \item Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})
}

}
